+++ src/libbu/color.c
@@ -180,14 +180,14 @@ int bu_str_to_rgb(char *str, unsigned char *rgb)
     }
 
     r = g = b = -1;
-    while (isspace(*str))
+    while (isspace((int)(*str)))
 	++str;
 
     if (*str == '#') {
 	if (strlen(++str) != 6)
 	    return 0;
 	num = sscanf(str, "%02x%02x%02x", (unsigned int *)&r, (unsigned int *)&g, (unsigned int *)&b);
-    } else if (isdigit(*str)) {
+    } else if (isdigit((int)(*str))) {
 	num = sscanf(str, "%d/%d/%d", &r, &g, &b);
 	if (num == 1) {
 	    num = sscanf(str, "%d %d %d", &r, &g, &b);
