+++ src/other/libregex/regex2.h
@@ -70,11 +70,11 @@
  * In state representations, an operator's bit is on to signify a state
  * immediately *preceding* "execution" of that operator.
  */
-typedef unsigned long sop;	/* strip operator */
-typedef long sopno;
+typedef size_t sop;	/* strip operator */
+typedef ssize_t sopno;
 #define	OPRMASK	0xf8000000
 #define	OPDMASK	0x07ffffff
-#define	OPSHIFT	((unsigned)27)
+#define	OPSHIFT	((size_t)27)
 #define	OP(n)	((n)&OPRMASK)
 #define	OPND(n)	((n)&OPDMASK)
 #define	SOP(op, opnd)	((op)|(opnd))
